requirements.txt
qrcode[pil]>=8.0,<9.0
opencv-python>=4.11,<5.0
numpy>=2.2,<3.0
pillow>=11.1,<12.0
